+++ db/database.json
@@ -118,7 +118,7 @@
       "orderDate": "07-23-2018",
       "shipDate": "07-30-2018",
       "arrivalDate": "08-01-2018",
-      "orderStatus": "Coating",
+      "orderStatus": "Edging",
       "remake": false,
       "frameName": "0154",
       "frameManufacturer": "Oakley",
@@ -152,8 +152,8 @@
       "orderDate": "07-22-2018",
       "shipDate": "07-29-2018",
       "arrivalDate": "07-31-2018",
-      "orderStatus": "Edging",
-      "remake": false,
+      "orderStatus": "Blocking",
+      "remake": true,
       "frameName": "Hampton",
       "frameManufacturer": "Garret Leight",
       "frameMaterial": "Zyl",
@@ -186,7 +186,7 @@
       "orderDate": "07-17-2018",
       "shipDate": "07-31-2018",
       "arrivalDate": "08-02-2018",
-      "orderStatus": "Blocking",
+      "orderStatus": "Shipped",
       "remake": false,
       "frameName": "Lemtosh",
       "frameManufacturer": "Moscot",
@@ -220,7 +220,7 @@
       "orderDate": "07-22-2018",
       "shipDate": "07-29-2018",
       "arrivalDate": "07-31-2018",
-      "orderStatus": "Edging",
+      "orderStatus": "Mounting",
       "remake": false,
       "frameName": "Zev",
       "frameManufacturer": "Moscot",
@@ -253,7 +253,7 @@
       "orderDate": "8/2/2018",
       "shipDate": "8/9/2018",
       "arrivalDate": "8/11/2018",
-      "orderStatus": "Polishing",
+      "orderStatus": "Blocking",
       "remake": false,
       "frameName": "Hampton",
       "frameManufacturer": "Garrett Leight",
